feat: lazy loading technically works
